+++ package.json
@@ -46,6 +46,8 @@
     "clsx": "^1.0.4",
     "draft-js": "^0.11.0",
     "history": "^4.7.2",
+    "i18next": "^19.7.0",
+    "i18next-xhr-backend": "^3.2.2",
     "immutable": "^4.0.0-rc.12",
     "is-promise": "^4.0.0",
     "js-cookie": "^2.2.0",
@@ -62,6 +64,7 @@
     "react-dom": "^16.8.6",
     "react-dropzone": "^10.1.5",
     "react-helmet": "^5.2.1",
+    "react-i18next": "^11.7.2",
     "react-markdown": "^4.0.8",
     "react-paginate": "^6.3.0",
     "react-perfect-scrollbar": "^1.5.3",
