+++ public/locales/en/auth.json
@@ -1,0 +1,10 @@
+{
+    "title": "Log In",
+    "sub_title": "Sign in on the internal platform",
+    "user_name": "User name",
+    "no_account": "Don't have an account? Click here!",
+    "password": "password",
+    "login_button": "Log in",
+    "loading": "loading..."
+  }
+
