+++ public/locales/en/top_bar.json
@@ -1,0 +1,3 @@
+{
+  "title": "Let's chat!"
+}
