+++ public/locales/kin/auth.json
@@ -1,0 +1,10 @@
+{
+    "title": "Injira",
+    "sub_title": "Shyiramo email n'umubare w'ibanga winjire",
+    "user_name": "emayili",
+    "no_account": "Nta konti ufite? kanda hano",
+    "password": "Ijambo ry'ibanga",
+    "login_button": "Injira",
+    "loading": "Tegereza..."
+  }
+
