+++ public/locales/kin/top_bar.json
@@ -1,0 +1,3 @@
+{
+  "title": "Tuganire!"
+}
