+++ src/layouts/Auth/styles/index.js
@@ -1,0 +1,1 @@
+export { default as useStyles } from './Styles';
