+++ src/views/Login/Login.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Link as RouterLink } from 'react-router-dom';
-import { makeStyles } from '@material-ui/styles';
 import {
   Card,
   CardContent,
@@ -11,74 +10,13 @@ import {
   Avatar
 } from '@material-ui/core';
 import LockIcon from '@material-ui/icons/Lock';
+import PropTypes from 'prop-types';
 
 import { Page } from 'components';
-import gradients from 'utils/gradients';
-import { LoginForm } from './components';
+import { LoginForm, useStyles } from './components';
 
-const useStyles = makeStyles(theme => ({
-  root: {
-    height: '100%',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: theme.spacing(6, 2)
-  },
-  card: {
-    width: theme.breakpoints.values.md,
-    maxWidth: '100%',
-    overflow: 'unset',
-    display: 'flex',
-    position: 'relative',
-    '& > *': {
-      flexGrow: 1,
-      flexBasis: '50%',
-      width: '50%'
-    }
-  },
-  content: {
-    padding: theme.spacing(8, 4, 3, 4)
-  },
-  media: {
-    borderTopRightRadius: 4,
-    borderBottomRightRadius: 4,
-    padding: theme.spacing(3),
-    color: theme.palette.white,
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'flex-end',
-    [theme.breakpoints.down('md')]: {
-      display: 'none'
-    }
-  },
-  icon: {
-    backgroundImage: gradients.green,
-    color: theme.palette.white,
-    borderRadius: theme.shape.borderRadius,
-    padding: theme.spacing(1),
-    position: 'absolute',
-    top: -32,
-    left: theme.spacing(3),
-    height: 64,
-    width: 64,
-    fontSize: 32
-  },
-  loginForm: {
-    marginTop: theme.spacing(3)
-  },
-  divider: {
-    margin: theme.spacing(2, 0)
-  },
-  person: {
-    marginTop: theme.spacing(2),
-    display: 'flex'
-  },
-  avatar: {
-    marginRight: theme.spacing(2)
-  }
-}));
-
-const Login = () => {
+const Login = (props) => {
+  const { t } = props;
   const classes = useStyles();
 
   return (
@@ -93,12 +31,12 @@ const Login = () => {
             gutterBottom
             variant="h3"
           >
-            Sign in
+            {t('auth:title')}
           </Typography>
           <Typography variant="subtitle2">
-            Sign in on the internal platform
+            {t('auth:sub_title')}
           </Typography>
-          <LoginForm className={classes.loginForm} />
+          <LoginForm className={classes.loginForm} t={t} />
           <Divider className={classes.divider} />
           <Link
             align="center"
@@ -108,7 +46,7 @@ const Login = () => {
             underline="always"
             variant="subtitle2"
           >
-            Don't have an account?
+            {t('auth:no_account')}
           </Link>
         </CardContent>
         <CardMedia
@@ -149,5 +87,10 @@ const Login = () => {
     </Page>
   );
 };
+
+Login.propTypes = {
+  t: PropTypes.func
+};
+
 
 export default Login;
