+++ src/views/Login/components/index.js
@@ -1,1 +1,2 @@
 export { default as LoginForm } from './LoginForm';
+export { default as useStyles } from './styles';
