+++ src/views/Login/components/styles/index.js
@@ -1,0 +1,1 @@
+export { default } from './Styles';
